pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [ "poetry-core>=2,<3" ]

[project]
name = "advattacks"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
  { name = "BoChenSHEN", email = "6goddddddd@gmail.com" },
]
requires-python = ">=3.11,<3.15"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "foolbox>=3.3.4,<4",
  "loguru>=0.7.3,<0.8",
  "pillow>=12,<13",
  "torch>=2.9.1,<3",
  "torchattacks>=3.5.1,<4",
  "torchvision>=0.24.1,<0.25",
  "tqdm>=4.67.1,<5",
  "transformers>=4.57.3,<5",
]

optional-dependencies.dev = [ "pre-commit>=4.5,<5", "ruff>=0.14.6,<0.15" ]
